8202579: Revisit VM_Version and VM_Version_ext for overlap and consolidation
