+++ src/hotspot/cpu/arm/vm_version_arm.hpp
@@ -106,6 +106,7 @@ class VM_Version: public Abstract_VM_Version {
 
   friend class VM_Version_StubGenerator;
 
+  static void initialize_cpu_information(void);
 };
 
 #endif // CPU_ARM_VM_VERSION_ARM_HPP
