+++ src/hotspot/cpu/s390/vm_version_s390.cpp
@@ -1490,3 +1490,19 @@ unsigned long VM_Version::z_SIGSEGV() {
  );
   return ZeroBuffer;
 }
+
+
+// get cpu information.
+void VM_Version::initialize_cpu_information(void) {
+  // do nothing if cpu info has been initialized
+  if (_initialized) {
+    return;
+  }
+
+  _no_of_cores  = os::processor_count();
+  _no_of_threads = _no_of_cores;
+  _no_of_sockets = _no_of_cores;
+  snprintf(_cpu_name, CPU_TYPE_DESC_BUF_SIZE, "s390 %s", VM_Version::get_model_string());
+  snprintf(_cpu_desc, CPU_DETAILED_DESC_BUF_SIZE, "s390 %s", features_string());
+  _initialized = true;
+}
