+++ src/hotspot/share/prims/jvm.cpp
@@ -3648,16 +3648,16 @@ JVM_END
 
 JVM_ENTRY_NO_ENV(jlong, JVM_GetRandomSeedForDumping())
   if (DumpSharedSpaces) {
-    const char* release = Abstract_VM_Version::vm_release();
-    const char* dbg_level = Abstract_VM_Version::jdk_debug_level();
+    const char* release = VM_Version::vm_release();
+    const char* dbg_level = VM_Version::jdk_debug_level();
     const char* version = VM_Version::internal_vm_info_string();
     jlong seed = (jlong)(java_lang_String::hash_code((const jbyte*)release, (int)strlen(release)) ^
                          java_lang_String::hash_code((const jbyte*)dbg_level, (int)strlen(dbg_level)) ^
                          java_lang_String::hash_code((const jbyte*)version, (int)strlen(version)));
-    seed += (jlong)Abstract_VM_Version::vm_major_version();
-    seed += (jlong)Abstract_VM_Version::vm_minor_version();
-    seed += (jlong)Abstract_VM_Version::vm_security_version();
-    seed += (jlong)Abstract_VM_Version::vm_patch_version();
+    seed += (jlong)VM_Version::vm_major_version();
+    seed += (jlong)VM_Version::vm_minor_version();
+    seed += (jlong)VM_Version::vm_security_version();
+    seed += (jlong)VM_Version::vm_patch_version();
     if (seed == 0) { // don't let this ever be zero.
       seed = 0x87654321;
     }
