+++ src/hotspot/share/prims/jvmtiExport.cpp
@@ -358,7 +358,7 @@ JvmtiExport::get_jvmti_interface(JavaVM *jvm, void **penv, jint version) {
       break;
     default:
       // Starting from 13 we do not care about minor version anymore
-      if (major < 13 || major > Abstract_VM_Version::vm_major_version()) {
+      if (major < 13 || major > VM_Version::vm_major_version()) {
         return JNI_EVERSION;  // unsupported major version number
       }
   }
