+++ src/hotspot/share/runtime/abstract_vm_version.hpp
@@ -186,6 +186,26 @@ class Abstract_VM_Version: AllStatic {
   constexpr static bool supports_stack_watermark_barrier() { return false; }
 
   static bool print_matching_lines_from_file(const char* filename, outputStream* st, const char* keywords_to_match[]);
+
+ protected:
+  // VM_Version statics
+  static const size_t      CPU_TYPE_DESC_BUF_SIZE = 256;
+  static const size_t      CPU_DETAILED_DESC_BUF_SIZE = 4096;
+
+  static int   _no_of_threads;
+  static int   _no_of_cores;
+  static int   _no_of_sockets;
+  static bool  _initialized;
+  static char  _cpu_name[CPU_TYPE_DESC_BUF_SIZE];
+  static char  _cpu_desc[CPU_DETAILED_DESC_BUF_SIZE];
+
+ public:
+  static int number_of_threads(void);
+  static int number_of_cores(void);
+  static int number_of_sockets(void);
+
+  static const char* cpu_name(void);
+  static const char* cpu_description(void);
 };
 
 #endif // SHARE_RUNTIME_ABSTRACT_VM_VERSION_HPP
